server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
langchain==0.0.340
tiktoken==0.5.1
numpy==1.24.3
requests==2.31.0
urllib3==2.0.7

# Google GenAI and Gemini (updated for direct audio support)
langchain-google-genai==1.0.10
google-generativeai>=0.7.0

# MongoDB Vector Database
pymongo==4.6.0
langchain-mongodb==0.1.3

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.9.0

# Image processing and OCR
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78

# File handling
Werkzeug==2.3.7

# WebSocket support
python-socketio[client]==5.9.0

# Testing
requests==2.31.0